remove duplication on results
