+++ app/presenters/events/scoreboards/results.rb
@@ -1,0 +1,9 @@
+module Events
+  module Scoreboards
+    module Results
+      def results
+        @results ||= Collection.new(event.event_tracks, params)
+      end
+    end
+  end
+end
